chore: using new endpoint to get metadata info and misc changes
