+++ src/components/TopBarNetworkComponent/TopBarNetworkConnect.tsx
@@ -1,92 +1,74 @@
-import { Button, Typography } from "@mui/material";
-import React, { useEffect } from "react";
-import LogoutIcon from "@mui/icons-material/Logout";
-import {
-  Actions,
-  BridgeConfig,
-  ExplorerContext,
-} from "../../types";
-import { getIconNamePerChainId, shortenAddress } from "../../utils/Helpers";
+import { Button, Typography } from "@mui/material"
+import React, { useEffect } from "react"
+import LogoutIcon from "@mui/icons-material/Logout"
+import { Actions, BridgeConfig, ExplorerContext, ExplorerContextState } from "../../types"
 
-import { useStyles } from "./styles";
+import { getIconNamePerChainId, shortenAddress } from "../../utils/transferHelpers"
+import { useStyles } from "./styles"
 
 type TopBarNetworkConnectProps = {
-  walletConnecting: boolean;
-  homeConfig: BridgeConfig | undefined;
-  address: string | undefined;
-  getAccount: ExplorerContext["getAccount"];
-  getChainId: ExplorerContext["getChainId"];
-  chainId: ExplorerContext["chainId"];
-  account: ExplorerContext["account"];
-  explorerContextDispatcher: React.Dispatch<Actions>;
-};
+  walletConnecting: boolean
+  homeConfig: BridgeConfig | undefined
+  address: string | undefined
+  getAccount: ExplorerContext["getAccount"]
+  getChainId: ExplorerContext["getChainId"]
+  chainId: ExplorerContext["chainId"]
+  account: ExplorerContext["account"]
+  explorerContextDispatcher: React.Dispatch<Actions>
+  explorerContextState?: ExplorerContextState
+}
 
-export default function TopBarNetworkConnect({
-  getAccount,
-  getChainId,
-  chainId,
-  account,
-  explorerContextDispatcher,
-}: TopBarNetworkConnectProps) {
-  const { classes } = useStyles();
-  const [localAddress, setLocalAddress] = React.useState<string | undefined>(
-    "",
-  );
-  const [isReady, setIsReady] = React.useState(false);
-  const [currentChainId, setCurrentChainId] = React.useState<
-    number | undefined
-  >(undefined);
+export default function TopBarNetworkConnect({ getAccount, getChainId, chainId, account, explorerContextDispatcher }: TopBarNetworkConnectProps) {
+  const { classes } = useStyles()
+  const [localAddress, setLocalAddress] = React.useState<string | undefined>("")
+  const [isReady, setIsReady] = React.useState(false)
+  const [currentChainId, setCurrentChainId] = React.useState<number | undefined>(undefined)
 
   const handleClickOpen = async () => {
-    const account = await getAccount();
-    const chainId = Number(await getChainId());
+    const account = await getAccount()
+    const chainId = Number(await getChainId())
 
-    setCurrentChainId(chainId);
+    setCurrentChainId(chainId)
 
-    setLocalAddress(account);
+    setLocalAddress(account)
 
     explorerContextDispatcher({
       type: "set_my_address",
       payload: account,
-    });
+    })
 
-    setIsReady(true);
-  };
+    setIsReady(true)
+  }
 
   const handleDisconnect = () => {
-    setCurrentChainId(undefined);
+    setCurrentChainId(undefined)
 
-    setLocalAddress("");
+    setLocalAddress("")
 
     explorerContextDispatcher({
       type: "set_my_address",
       payload: undefined,
-    });
+    })
 
-    setIsReady(false);
-  };
+    setIsReady(false)
+  }
 
   useEffect(() => {
     if (chainId !== undefined && chainId !== currentChainId) {
-      setCurrentChainId(chainId);
+      setCurrentChainId(chainId)
     }
 
     if (account !== undefined && account !== localAddress) {
-      setLocalAddress(account);
+      setLocalAddress(account)
     }
-  }, [chainId, account]);
+  }, [chainId, account])
 
   return (
     <>
       <section className={classes.state}>
         {window.ethereum !== undefined &&
           (!isReady ? (
-            <Button
-              fullWidth
-              variant="contained"
-              onClick={handleClickOpen}
-              sx={{ px: 3, fontSize: 18 }}
-            >
+            <Button fullWidth variant="contained" onClick={handleClickOpen} sx={{ px: 3, fontSize: 18 }}>
               Connect Wallet
             </Button>
           ) : (
@@ -102,13 +84,7 @@ export default function TopBarNetworkConnect({
               >
                 <Typography variant="h6" className={classes.address}>
                   {currentChainId !== undefined && (
-                    <img
-                      src={`/assets/icons/${getIconNamePerChainId(
-                        currentChainId,
-                      )}`}
-                      alt={"native token icon"}
-                      className={classes.indicator}
-                    />
+                    <img src={`/assets/icons/${getIconNamePerChainId(currentChainId)}`} alt={"native token icon"} className={classes.indicator} />
                   )}
                 </Typography>
                 <div className={classes.accountInfo}>
@@ -116,10 +92,7 @@ export default function TopBarNetworkConnect({
                     {shortenAddress(localAddress!)}
                   </Typography>
                   <div className={classes.logoutContainer}>
-                    <div
-                      className={classes.logoutIconContainer}
-                      onClick={handleDisconnect}
-                    >
+                    <div className={classes.logoutIconContainer} onClick={handleDisconnect}>
                       <LogoutIcon />
                     </div>
                   </div>
@@ -129,5 +102,5 @@ export default function TopBarNetworkConnect({
           ))}
       </section>
     </>
-  );
+  )
 }
