+++ src/context/reducer.ts
@@ -44,6 +44,18 @@ export function reducer(state: ExplorerContextState, action: Actions): ExplorerC
         domainMetadata: action.payload,
       }
     }
+    case "set_resources_per_page": {
+      return {
+        ...state,
+        resourcesPerPage: action.payload,
+      }
+    }
+    case "set_source_domains_ids": {
+      return {
+        ...state,
+        sourceDomainsIds: action.payload,
+      }
+    }
     default:
       return state
   }
